+++ src/main/kotlin/com/atlassian/performance/tools/jiraactions/api/measure/ActionMeter.kt
@@ -12,6 +12,7 @@ import com.atlassian.performance.tools.jiraactions.api.w3c.W3cPerformanceTimelin
 import java.time.Clock
 import java.time.Duration
 import java.util.*
+import java.util.function.Predicate
 import javax.json.JsonObject
 
 /**
@@ -21,16 +22,35 @@ import javax.json.JsonObject
  * @param output writes the metrics
  * @param clock measures latencies
  * @param w3cPerformanceTimeline records W3C performance entries
+ * @param drilldownCondition condition based on ActionMetric upon which drilldown is performed
  */
-class ActionMeter(
+class ActionMeter private constructor(
     private val virtualUser: UUID,
     private val output: ActionMetricOutput,
     private val clock: Clock,
-    private val w3cPerformanceTimeline: W3cPerformanceTimeline
-) {
+    private val w3cPerformanceTimeline: W3cPerformanceTimeline,
+    private val drilldownCondition: Predicate<ActionMetric>
+)
+{
 
     @Deprecated(
-        message = "Use the primary constructor"
+        message = "Use ActionMeter.Builder instead"
+    )
+    constructor(
+        virtualUser: UUID,
+        output: ActionMetricOutput,
+        clock: Clock,
+        w3cPerformanceTimeline: W3cPerformanceTimeline
+    ) : this (
+        virtualUser = virtualUser,
+        output = output,
+        clock = clock,
+        w3cPerformanceTimeline = w3cPerformanceTimeline,
+        drilldownCondition = Predicate { true }
+    )
+
+    @Deprecated(
+        message = "Use ActionMeter.builder instead"
     )
     constructor(
         virtualUser: UUID,
@@ -40,7 +60,8 @@ class ActionMeter(
         virtualUser = virtualUser,
         output = output,
         clock = clock,
-        w3cPerformanceTimeline = DisabledW3cPerformanceTimeline()
+        w3cPerformanceTimeline = DisabledW3cPerformanceTimeline(),
+        drilldownCondition = Predicate { false }
     )
 
     /**
@@ -76,50 +97,80 @@ class ActionMeter(
         observation: (T) -> JsonObject?
     ): T {
         val start = clock.instant()
+        var actionResult: ActionResult = ActionResult.OK
+        var result: T? = null
+
         try {
-            val result = action()
-            output.write(
-                ActionMetric.Builder(
-                    label = key.label,
-                    result = ActionResult.OK,
-                    start = start,
-                    duration = Duration.between(start, clock.instant())
-                )
-                    .virtualUser(virtualUser)
-                    .observation(observation(result))
-                    .drilldown(w3cPerformanceTimeline.record())
-                    .build()
-            )
+            result = action()
             return result
         } catch (e: Exception) {
-            val result = if (e.representsInterrupt()) {
+            actionResult = if (e.representsInterrupt()) {
                 ActionResult.INTERRUPTED
             } else {
                 ActionResult.ERROR
             }
-            output.write(
-                ActionMetric.Builder(
-                    label = key.label,
-                    result = result,
-                    start = start,
-                    duration = Duration.between(start, clock.instant())
-                )
-                    .virtualUser(virtualUser)
+            throw Exception("Action '${key.label}' $actionResult", e)
+        } finally {
+            val metricBuilder = ActionMetric.Builder(
+                label = key.label,
+                result = actionResult,
+                start = start,
+                duration = Duration.between(start, clock.instant())
+            )
+                .virtualUser(virtualUser)
+
+            result?.let { metricBuilder.observation(observation(result)) }
+
+            var metric = metricBuilder.build()
+
+            if (drilldownCondition.test(metric)) {
+                metric = ActionMetric.Builder(metric)
                     .drilldown(w3cPerformanceTimeline.record())
                     .build()
-            )
-            throw Exception("Action '${key.label}' $result", e)
+            }
+            output.write(metric)
         }
     }
 
+    @Deprecated( message = "Use builder instead")
     fun withW3cPerformanceTimeline(
         w3cPerformanceTimeline: W3cPerformanceTimeline
     ): ActionMeter = ActionMeter(
         virtualUser = virtualUser,
         output = output,
         clock = clock,
-        w3cPerformanceTimeline = w3cPerformanceTimeline
+        w3cPerformanceTimeline = w3cPerformanceTimeline,
+        drilldownCondition = Predicate { true }
     )
+
+    class Builder(
+        private val virtualUser: UUID,
+        private val output: ActionMetricOutput,
+        private val clock: Clock
+    ){
+        private var w3cPerformanceTimeline: W3cPerformanceTimeline = DisabledW3cPerformanceTimeline()
+        private var drilldownCondition: Predicate<ActionMetric> = Predicate { true }
+
+        constructor(meter: ActionMeter): this(
+            meter.virtualUser,
+            meter.output,
+            meter.clock
+        ){
+            w3cPerformanceTimeline = meter.w3cPerformanceTimeline
+            drilldownCondition = meter.drilldownCondition
+        }
+
+        fun performanceTimeline(w3cPerformanceTimeline: W3cPerformanceTimeline) = apply { this.w3cPerformanceTimeline = w3cPerformanceTimeline }
+        fun drilldownCondition(drilldownCondition: Predicate<ActionMetric>) = apply { this.drilldownCondition = drilldownCondition }
+
+        fun build() : ActionMeter = ActionMeter(
+            virtualUser = virtualUser,
+            output = output,
+            clock = clock,
+            w3cPerformanceTimeline = w3cPerformanceTimeline,
+            drilldownCondition = drilldownCondition
+        )
+    }
 }
 
 @Deprecated("This is an internal data structure used by ActionMeter. Use ActionMeter public methods instead.")